GH-72 - Collaboration Awareness
Show collab awareness status as icons just left of the publish button

If the user is known their initial is displayed.
If the user is not known a generic 'user' icon is displayed.

When hovering over the icon the full name is shown as a tooltip.

When collab is disconnected disable the editor until it reconnects.

Collab UX Polish

* Simplified integration between prose/index and da-title.
* prose/index only responsible for data, da-title is presentation.
* Turned user icon into clickable initial pill. Clicking shows full name popover.
* Cloud is also clickable for more detailed status