New Playwright tests for the versions functionality.
Also updated the other Playwright test to not have to clean up their
temporary files in a finally block any more, but rather clean these up
together in a separate 'delete' test that cleans up matching files if
they are older than 2 hours. It does this cleanup through the UI so it
also serves as a test for delete.