Playwright tests for ACL functionality
These tests log into IMS as a 'DA Test User' and use the
https://github.com/da-testautomation/acltest site/repository for their
tests.

Also updated the collab tests to use an actually logged in user rather
than 2 Anonymous users.

NOTE THAT WHEN LOGGED IN TESTS THAT DEPEND ON THAT ALWAYS USE THE
https://da.live DOMAIN. THIS IS BECAUSE THATS WHERE IMS LOGS US IN!

Also deleted 2 empty .css files that cause the linter to fail.