fix: use document GUID as prosemirror reference to avoid updating deleted documents
Store the current prosemirror document in the YDoc under prosemirror-${guid} instead of the previous prosemirror key.
A new prosemirror document gets a new GUID, even if it was stored in the same location as where a previous document used to live. This means that if there was an old editor open somewhere, that is editing a document that has since been deleted, these edits are then ignored.

The property prosemirror-guids contains an array of GUIDs with timestamps. Only the newest timestamp is active, any older guids on that array are not active any more. This array is often scrubbed, do deleted documents on the same location often don't appear in the prosemirror-guids at all any more.

When creating a fresh new document in the editor, create a new GUID for it, using crypto.randomUUID()
When editing an existing document, the GUID of that document is returned from the HEAD request in the x-da-id header. So we can set that to the prosemirror-${guid} in that case.

Requires existing open browser DA editors to be refreshed.